Extracted nested classes from subscription manager
