+++ Src/SwqlStudio/SwqlStudio.csproj
@@ -226,6 +226,8 @@
     <Compile Include="ServerList.cs" />
     <Compile Include="ServerType.cs" />
     <Compile Include="Subscription.cs" />
+    <Compile Include="Subscriptions\SubscriptionCallbacks.cs" />
+    <Compile Include="Subscriptions\SubscriptionInfo.cs" />
     <Compile Include="Subscriptions\SubscriptionManager.cs" />
     <Compile Include="Subscriptions\SubscriptionServiceHost.cs" />
     <Compile Include="SubscriptionTab.cs">
